[FEAT/#352] 토글 컴포넌트화
